int time and dt, production/construction queue removed, added to order system. Added OnQueue, OnCancel actions for orders, Started new construction method where resources are dumped into temporary construction storage (in the structure) and units who are constructing the building (multiple allowed) take from the pile (or add to it if they have needed resources). Removed all references to Time.time (unity) and added StepManager.GameTime. Wrote function to calculate costs per game step of construction and account for rounding issues. Resources are consumed divided evenly over construction time, with a lack of any preventing the continuation of progress. The first construction tick requires 1 of each resource type, and the last tick requires 0. Tried to fix collision jitter but putting that on hold now, unity physics engine is non-deterministic so we will need to write our own with anything we need to simulate deterministically (for now just need a quad tree for collision detection). StartNewStructure no longer takes quantity, consolidated moveUnit functions, fixed network step bug in main game loop. Also did some other minor refactoring and renaming.

+++ Assets/Scripts/Shared/Abilities/Ability.cs
@@ -5,8 +5,8 @@ public class Ability : MyMonoBehaviour {
 
     public RTSGameObject target;
     public Vector3 targetPosition;
-    public float range;
-    public float cooldown;
+    public int range;
+    public int cooldown;
 
     public override string ToString()
     {
@@ -38,8 +38,8 @@ public class Ability : MyMonoBehaviour {
         Ability ability = AbilityFactory.GetAbilityFromString(abilityComponents[0]);
         ability.target = GameObject.Find(abilityComponents[1]).GetComponent<RTSGameObject>();
         ability.targetPosition = targetPosition;
-        ability.range = float.Parse(abilityComponents[3]);
-        ability.cooldown = float.Parse(abilityComponents[4]);
+        ability.range = int.Parse(abilityComponents[3]);
+        ability.cooldown = int.Parse(abilityComponents[4]);
 
         return ability;
     }
